requirements.txt
pandas>=1.5.0
ofxparse>=0.20
requests>=2.28.0
python-dateutil>=2.8.2
pytz>=2022.1
SQLAlchemy>=1.4.0
numpy>=1.23.0
pytest>=7.1.0
python-dotenv>=0.20.0
flask>=2.1.0
oandapyV20>=0.7.0
qlib>=0.6.9
rich>=12.0.0